Move _handle_line() to _marks.py
