+++ mplaltair/_convert.py
@@ -1,5 +1,5 @@
-import matplotlib.dates as mdates
-from ._data import _locate_channel_data, _locate_channel_dtype, _convert_to_mpl_date, _normalize_data
+from ._data import _locate_channel_data, _locate_channel_dtype
+
 
 def _allowed_ranged_marks(enc_channel, mark):
     """TODO: DOCS
@@ -93,12 +93,6 @@ _mappings = {
     'stroke': _process_stroke,
 }
 
-# _line_mapping = {
-#     'x': lambda d: ('x', d),
-#     'y': lambda d: ('y', d),
-#     'args': lambda x, y: [x, y]
-# }
-
 def _convert(chart):
     """Convert an altair encoding to a Matplotlib figure
 
@@ -131,12 +125,4 @@ def _convert(chart):
 
         mapping[_mappings[channel](dtype, data)[0]] = _mappings[channel](dtype, data)[1]
 
-    #     if chart.mark in ['point', 'circle', 'square']:
-    #         mapping[_mappings[channel](dtype, data)[0]] = _mappings[channel](dtype, data)[1]
-    #     elif chart.mark == 'line' and channel in ['x', 'y']:
-    #         mapping[_line_mapping[channel](data)[0]] = _line_mapping[channel](data)[1]
-    #
-    # if chart.mark == 'line':
-    #     mapping['args'] = _line_mapping['args'](mapping['x'], mapping['y'])  # plot() doesn't take kwargs for x and y
-    
     return mapping
