+++ mplaltair/tests/test_convert.py
@@ -2,7 +2,6 @@ import pytest
 
 import altair as alt
 import pandas as pd
-import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
 
 from mplaltair import convert
